# [Nonlinear] fix performance of evaluating univariate operators