--- a/test/Nonlinear/Nonlinear.jl
+++ b/test/Nonlinear/Nonlinear.jl
@@ -365,6 +365,7 @@ function test_eval_univariate_function()
         (:-, 1.0, -1.0),
         (:abs, -1.1, 1.1),
         (:abs, 1.1, 1.1),
+        (:sin, 1.1, sin(1.1)),
     ]
         id = r.univariate_operator_to_id[op]
         @test Nonlinear.eval_univariate_function(r, op, x) == y
@@ -380,6 +381,7 @@ function test_eval_univariate_gradient()
         (:-, 1.2, -1.0),
         (:abs, -1.1, -1.0),
         (:abs, 1.1, 1.0),
+        (:sin, 1.1, cos(1.1)),
     ]
         id = r.univariate_operator_to_id[op]
         @test Nonlinear.eval_univariate_gradient(r, op, x) == y
@@ -395,6 +397,7 @@ function test_eval_univariate_function_and_gradient()
         (:-, 1.2, (-1.2, -1.0)),
         (:abs, -1.1, (1.1, -1.0)),
         (:abs, 1.1, (1.1, 1.0)),
+        (:sin, 1.1, (sin(1.1), cos(1.1))),
     ]
         id = r.univariate_operator_to_id[op]
         @test Nonlinear.eval_univariate_function_and_gradient(r, op, x) == y