With the energy modeling benchmarking that @joaquimg has been doing, we've been talking about how small things never show up in a flame graph until you look at their aggregate across multiple call sites. 

One thing that shows up is `getindex`

<img width="366" alt="image" src="https://github.com/user-attachments/assets/c13bf398-11bb-4e5f-8c13-47acce1acce7" />

With the only giveaway being the `ht_keyindex(::Dict{Symbol,Int}, ::Symbol)` method.

It's because we're calling `operators.univariate_operators[node.index]` in multiple places, which is a `Dict` lookup.

These calls don't show much because of inlining etc. Even in the code view, you're kinda like "call is expensive. Sure"

<img width="1394" alt="image" src="https://github.com/user-attachments/assets/d6be3087-38fc-46fa-b00c-f8b66f57ca07" />

## Before

```Julia
julia> @benchmark(
           MOI.Nonlinear.ReverseAD._reverse_mode($(evaluator.backend), x),
           setup=(x=rand(num_variables(model))),
       )
BenchmarkTools.Trial: 44 samples with 1 evaluation per sample.
 Range (min … max):  112.457 ms … 132.677 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     113.303 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   115.478 ms ±   4.641 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

   █▂                                                            
  ▇███▆▃▁▄▁▁▁▁▁▃▁▁▁▁▁▃▁▁▁▁▃▃▄▃▁▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▁▁▁▁▁▃ ▁
  112 ms           Histogram: frequency by time          133 ms <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

## After

```Julia
julia> @benchmark(
           MOI.Nonlinear.ReverseAD._reverse_mode($(evaluator.backend), x),
           setup=(x=rand(num_variables(model))),
       )
BenchmarkTools.Trial: 48 samples with 1 evaluation per sample.
 Range (min … max):  102.579 ms … 115.444 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     104.459 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   105.355 ms ±   2.961 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

       ███▆█ ▁                                                   
  ▄▁▁▄▄█████▇█▄▇▄▄▁▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▄▁▁▄▁▄▁▁▁▁▄ ▁
  103 ms           Histogram: frequency by time          115 ms <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

## Script

```Julia
using JuMP, PowerModels, BenchmarkTools
case = "test/data/pglib_opf_case13659_pegase.m"
backend = MOI.Nonlinear.SparseReverseMode()
begin
    pm = PowerModels.instantiate_model(
        case,
        PowerModels.ACPPowerModel,
        PowerModels.build_opf,
    );
    model = pm.model
    nlp = MOI.Nonlinear.Model()
    for (F, S) in list_of_constraint_types(model)
        if F != NonlinearExpr
            continue
        end
        for ci in all_constraints(model, F, S)
            o = constraint_object(ci)
            MOI.Nonlinear.add_constraint(nlp, o.func, o.set)
        end
    end
    evaluator = MOI.Nonlinear.Evaluator(
        nlp,
        backend,
        index.(all_variables(model)),
    )
end;
MOI.initialize(evaluator, [:Grad, :Jac, :Hess])
```